set up initial View html files
