+++ src/main/java/com/example/movieGame/RunGame.java
@@ -1,0 +1,12 @@
+package com.example.movieGame;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RunGame {
+    public static void main(String[] args) {
+        SpringApplication.run(RunGame.class, args);
+    }
+}
+
